IDVA6-1274 test files moved to the correct directory
