+++ test/src/routers/controllers/removeUserCheckDetailsController.test.ts
@@ -1,13 +1,13 @@
-import mocks from "../../mocks/all.middleware.mock";
+import mocks from "../../../mocks/all.middleware.mock";
 import supertest from "supertest";
-import app from "../../../src/app";
-import * as en from "../../../src/locales/en/translation/remove-member.json";
-import * as enCommon from "../../../src/locales/en/translation/common.json";
-import * as constants from "../../../src/lib/constants";
+import app from "../../../../src/app";
+import * as en from "../../../../src/locales/en/translation/remove-member.json";
+import * as enCommon from "../../../../src/locales/en/translation/common.json";
+import * as constants from "../../../../src/lib/constants";
 import { Session } from "@companieshouse/node-session-handler";
 import { Request, Response, NextFunction } from "express";
-import { setExtraData } from "../../../src/lib/utils/sessionUtils";
-import { Membership } from "../../../src/types/membership";
+import { setExtraData } from "../../../../src/lib/utils/sessionUtils";
+import { Membership } from "../../../../src/types/membership";
 
 const router = supertest(app);
