+++ test/src/routers/controllers/removeUserSuccessController.test.ts
@@ -1,12 +1,12 @@
-import mocks from "../../mocks/all.middleware.mock";
+import mocks from "../../../mocks/all.middleware.mock";
 import supertest from "supertest";
-import app from "../../../src/app";
-import * as en from "../../../src/locales/en/translation/confirmation-member-removed.json";
-import * as enCommon from "../../../src/locales/en/translation/common.json";
-import * as constants from "../../../src/lib/constants";
+import app from "../../../../src/app";
+import * as en from "../../../../src/locales/en/translation/confirmation-member-removed.json";
+import * as enCommon from "../../../../src/locales/en/translation/common.json";
+import * as constants from "../../../../src/lib/constants";
 import { Session } from "@companieshouse/node-session-handler";
 import { Request, Response, NextFunction } from "express";
-import { setExtraData } from "../../../src/lib/utils/sessionUtils";
+import { setExtraData } from "../../../../src/lib/utils/sessionUtils";
 
 const router = supertest(app);
